Move PySNMP imports to a separate pysnmp_types module (#6013)
